+++ packages/client/src/pages/Forum/components/ForumMetaData.tsx
@@ -1,0 +1,17 @@
+import { Helmet } from 'react-helmet'
+
+type IProps = {
+  title: string
+}
+
+export default function ForumMetaData({ title }: IProps) {
+  return (
+    <>
+      <Helmet>
+        <meta charSet="utf-8" />
+        <title>{title}</title>
+        <meta name="description" content={title} />
+      </Helmet>
+    </>
+  )
+}
