+++ packages/client/src/pages/Forum/components/ForumTextareaComment/styles.module.scss
@@ -1,0 +1,45 @@
+.textarea_container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  .textarea {
+    width: 56.25rem;
+    height: 12.125rem;
+    padding: 1.25rem;
+    margin-bottom: 1.25rem;
+    resize: none;
+    outline: none;
+  }
+
+  .buttons_container {
+    position: relative;
+    width: 56.25rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    .emoji_picker {
+      position: absolute;
+      bottom: 3rem;
+      left: 0;
+    }
+
+    .emoji_button {
+      width: 4.25rem;
+      height: 2.5rem;
+      background-image: url('/svg/smile.svg');
+      background-position: center;
+      background-repeat: no-repeat;
+      background-color: var(--primary-color-light);
+    }
+
+    .send_button {
+      width: 6.625rem;
+      height: 2.5rem;
+      background-color: var(--primary-color-light);
+      color: var(--background-color);
+    }
+  }
+}
