+++ packages/client/src/pages/Forum/components/ForumTopicTable/index.tsx
@@ -1,18 +1,12 @@
 import { ChangeEvent, useEffect, useMemo, useState, useTransition } from 'react'
 import styles from './styles.module.scss'
-import { Avatar, Button, TablePagination, Typography } from '@mui/material'
-import { Helmet } from 'react-helmet'
+import { TablePagination } from '@mui/material'
 import { useParams } from 'react-router-dom'
-import EmojiPicker from '@emoji-mart/react'
-import emojiData from '@emoji-mart/data'
 import { ForumTopicData } from '../constants'
-import {
-  createComment,
-  getTopicComments,
-  getTopicInfo,
-} from '../../../../api/basic/forum'
-import { useAppSelector } from '../../../../hooks/reduxHooks'
-import ForumTopicCommentTable from '../ForumTopicCommentTable'
+import { getTopicComments, getTopicInfo } from '../../../../api/basic/forum'
+import ForumMetaData from '../ForumMetaData'
+import ForumTextareaComment from '../ForumTextareaComment'
+import ForumTopicTableRow from './ForumTopicTableRow'
 
 export type ForumTopicTableRowDataType = {
   creator: JSX.Element
@@ -21,47 +15,20 @@ export type ForumTopicTableRowDataType = {
   replies_count: number
 }
 
-function getTableRowsData(
-  data: ForumTopicData[]
-): ForumTopicTableRowDataType[] {
-  return (
-    data.map(item => ({
-      ...item,
-      creator: (
-        <div className={styles.user}>
-          <Avatar
-            className={styles.avatar}
-            alt={item.creator.login}
-            src={item.creator.avatar}
-          />
-          <Typography variant="body1" fontSize="0.75rem">
-            {item.creator.login}
-          </Typography>
-        </div>
-      ),
-    })) ?? []
-  )
-}
-
 export default function ForumTopicTable() {
   const [isPending, startTransition] = useTransition()
   const { topicId } = useParams()
   const [page, setPage] = useState(0)
   const [rowsPerPage, setRowsPerPage] = useState(2)
-  const [textareaText, setTextareaText] = useState('')
-  const [showEmojiPicker, setShowEmojiPicker] = useState(false)
   const [commentsAmount, setCommentsAmount] = useState(0)
   const [topicName, setTopicName] = useState('')
   const [data, setData] = useState<ForumTopicData[]>([])
 
-  const user = useAppSelector(state => state.user.user)
-
   function setComments() {
-    return getTopicComments(
-      Number(topicId),
-      page * rowsPerPage,
-      rowsPerPage
-    ).then(data => {
+    return getTopicComments(Number(topicId), {
+      offset: page * rowsPerPage,
+      limit: rowsPerPage,
+    }).then(data => {
       setData(data)
     })
   }
@@ -77,18 +44,7 @@ export default function ForumTopicTable() {
     setComments()
   }, [page, rowsPerPage, topicId])
 
-  const tableRows = useMemo(
-    () =>
-      getTableRowsData(data).map(row => (
-        <ForumTopicCommentTable
-          key={row.id}
-          commentId={row.id}
-          row={row}
-          nestingLevel={0}
-        />
-      )),
-    [data]
-  )
+  const tableRows = useMemo(() => <ForumTopicTableRow data={data} />, [data])
 
   const handleChangePage = (event: unknown, newPage: number) => {
     startTransition(() => {
@@ -103,41 +59,9 @@ export default function ForumTopicTable() {
     })
   }
 
-  const textareaChangeHandle = (event: ChangeEvent<HTMLTextAreaElement>) => {
-    setTextareaText(event.target.value)
-  }
-
-  const showEmojiButtonHandle = () => {
-    setShowEmojiPicker(!showEmojiPicker)
-  }
-
-  const onEmojiSelectHandle = ({ native }: { native: string }) => {
-    setTextareaText(`${textareaText}${native}`)
-    setShowEmojiPicker(false)
-  }
-
-  const handleCreateComment = () => {
-    createComment({
-      topicId: Number(topicId),
-      content: textareaText,
-      creator: {
-        id: user.id,
-        login: user.login,
-        avatar: user.avatar,
-      },
-    }).then(() => {
-      setTextareaText('')
-      setComments().then(() => setCommentsAmount(commentsAmount + 1))
-    })
-  }
-
   return (
     <>
-      <Helmet>
-        <meta charSet="utf-8" />
-        <title>{topicName}</title>
-        <meta name="description" content={topicName} />
-      </Helmet>
+      <ForumMetaData title={topicName} />
       <div>{tableRows}</div>
       <TablePagination
         className={styles.pagination}
@@ -149,34 +73,11 @@ export default function ForumTopicTable() {
         onPageChange={handleChangePage}
         onRowsPerPageChange={handleChangeRowsPerPage}
       />
-      <div className={styles.textarea_container}>
-        <textarea
-          className={styles.textarea}
-          value={textareaText}
-          onChange={textareaChangeHandle}
-        />
-        <div className={styles.buttons_container}>
-          {showEmojiPicker && (
-            <div className={styles.emoji_picker}>
-              <EmojiPicker
-                data={emojiData}
-                onEmojiSelect={onEmojiSelectHandle}
-              />
-            </div>
-          )}
-          <Button
-            variant="contained"
-            className={styles.emoji_button}
-            onClick={showEmojiButtonHandle}
-          />
-          <Button
-            onClick={handleCreateComment}
-            variant="contained"
-            className={styles.send_button}>
-            SEND
-          </Button>
-        </div>
-      </div>
+      <ForumTextareaComment
+        callback={() =>
+          setComments().then(() => setCommentsAmount(commentsAmount + 1))
+        }
+      />
     </>
   )
 }
