+++ packages/client/src/pages/Forum/index.tsx
@@ -3,7 +3,7 @@ import { clsx } from 'clsx'
 import styles from './styles.module.scss'
 import PreviousPageBtn from '../../components/PreviousPageBtn'
 import { useParams } from 'react-router-dom'
-import ForumSubjectTable from './components/ForumSubjectTable'
+import ForumSubjectTable from './components/ForumTable'
 import { useEffect, useState } from 'react'
 import ForumTopicTable from './components/ForumTopicTable'
 import { useLoggedInUser } from '../../hooks/useLoggedInUser'
